Audit Feedback: uint256/uint16 type mismatch
**fixed**: variable is now uint16 everywhere.  did not follow recommendation of uint256 everywhere since it is not reasonable for this number to be larger than uint16, and client-side tooling is potentially simplified when the type is less than 52-bits (fits precisely inside of a double precision floating point number)

+++ contracts/source/recoverable-wallet.sol
@@ -71,7 +71,7 @@ contract Ownable {
 }
 
 contract RecoverableWallet is Ownable, Erc777TokensRecipient {
-	event RecoveryAddressAdded(address indexed newRecoverer, uint256 recoveryDelayInDays);
+	event RecoveryAddressAdded(address indexed newRecoverer, uint16 recoveryDelayInDays);
 	event RecoveryAddressRemoved(address indexed oldRecoverer);
 	event RecoveryStarted(address indexed newOwner);
 	event RecoveryCancelled();
